requirements.txt
# Core dependencies
streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
# Data visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Natural language processing
nltk>=3.8.0
textblob>=0.17.0

# Machine learning (optional for advanced features)
scikit-learn>=1.3.0

# Web scraping and APIs (for content enrichment)
beautifulsoup4>=4.12.0

# Database support
# sqlite3 is built-in with Python

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Environment management
python-dotenv>=1.0.0

# Date and time utilities
python-dateutil>=2.8.0

# JSON handling
jsonschema>=4.19.0

# File handling
Pillow>=10.0.0

# Progress bars
tqdm>=4.66.0

# Configuration management
pyyaml>=6.0.0

# Logging
loguru>=0.7.0